Debug control
Funny bug :
It is about the functions __getitem__ and __setitem__ of the FortranDerivedTypeArray class, at the very bottom of the fortrantype.py file.
Since python indexing is 0-based and fortran 1-based, a conversion has to be done. However, in the original code, the value of the argument is modified, instead of calling fortran with arg+1. If the argument is passed by reference, the former would modify  the calling index, which is wrong. However, I guess that the issue never presented itself because, when going through fortran, scalar arguments without intent are treated as intent in. But, if the index is declared as an array of length=1, it will be intent(inout), raising the problem.

I updated my fixed_1D_derived_type_array_argument example to test this: with the current code (call i+1) it works, but with the old code (i+=1), and "i" declared as bumpy.ones(1), it fails.