DerivedTypes: Enabling `pointer` arrays and fixing bug in tests
The handling of `pointer` arrays in derived types was disabled, but
seems to work once enabled (similar to `allocatable` arrays).

However, `test_return_dertype_pointer_arrays` in `derivedtypes` was
failing once the values of the pointer array are being checked.
This was due to the way the array was allocated (the pointer was
associated with a variable local to subroutine, which became
"undefined (or even unallocated) once we leave that subroutine,
causing undefined memory being given to the `numpy.array`.